requirements.txt
torch >= 1.0.0
torchvision
tqdm >= 4.65.0
h5py >= 3.7.0
numpy >= 1.11
pandas >= 0.20.3
scikit-learn >= 0.19.0
scipy >= 0.17.0
pillow >= 6.2.0
future